+++ test/unit.jl
@@ -20,6 +20,173 @@
         @test action.internal_transitions[1].by == Point(50, 50)
         Javis.compute_transformation!(action, video, 100)
         @test action.internal_transitions[1].by == Point(100, 100)
+
+        # with easing function
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, sineio(), Translation(Point(1, 1), Point(100, 100)))
+        anim = Animation([0.0, 1.0], [1.0, 100.0], [sineio()])
+        m = 49 / 99
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalTranslation(O))
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].by == Point(1, 1)
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].by == Point(at(anim, m), at(anim, m))
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].by == Point(100, 100)
+
+        # with animation function
+        anim_01 = Animation([0.0, 1.0], [0.0, 1.0], [sineio()])
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, anim_01, Translation(Point(1, 1), Point(100, 100)))
+        anim = Animation([0.0, 1.0], [1.0, 100.0], [sineio()])
+        m = 49 / 99
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalTranslation(O))
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].by == Point(1, 1)
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].by == Point(at(anim, m), at(anim, m))
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].by == Point(100, 100)
+    end
+
+    @testset "translation subaction" begin
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(
+            1:100,
+            () -> 1;
+            subactions = [SubAction(1:100, Translation(Point(1, 1), Point(100, 100)))],
+        )
+        # needs internal translation as well
+        push!(action.subactions[1].internal_transitions, Javis.InternalTranslation(O))
+        Javis.compute_transformation!(action.subactions[1], video, 1)
+        @test action.subactions[1].internal_transitions[1].by == Point(1, 1)
+        Javis.compute_transformation!(action.subactions[1], video, 50)
+        @test action.subactions[1].internal_transitions[1].by == Point(50, 50)
+        Javis.compute_transformation!(action.subactions[1], video, 100)
+        @test action.subactions[1].internal_transitions[1].by == Point(100, 100)
+
+        ######## testing with easing function ##############
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(
+            1:100,
+            () -> 1;
+            subactions = [SubAction(
+                1:100,
+                polyout(10),
+                Translation(Point(1, 1), Point(100, 100)),
+            )],
+        )
+        anim = Animation([0.0, 1.0], [1.0, 100.0], [polyout(10)])
+        m = 49 / 99
+        # needs internal translation as well
+        push!(action.subactions[1].internal_transitions, Javis.InternalTranslation(O))
+        Javis.compute_transformation!(action.subactions[1], video, 1)
+        @test action.subactions[1].internal_transitions[1].by == Point(1, 1)
+        Javis.compute_transformation!(action.subactions[1], video, 50)
+        @test action.subactions[1].internal_transitions[1].by ==
+              Point(at(anim, m), at(anim, m))
+        Javis.compute_transformation!(action.subactions[1], video, 100)
+        @test action.subactions[1].internal_transitions[1].by == Point(100, 100)
+
+        ######## testing with animation ##############
+        video = Video(500, 500)
+        anim_01 = Animation([0.0, 1.0], [0.0, 1.0], [polyin(10)])
+        # dummy action doesn't need a real function
+        action = Action(
+            1:100,
+            () -> 1;
+            subactions = [SubAction(
+                1:100,
+                anim_01,
+                Translation(Point(1, 1), Point(100, 100)),
+            )],
+        )
+        anim = Animation([0.0, 1.0], [1.0, 100.0], [polyin(10)])
+        m = 49 / 99
+        # needs internal translation as well
+        push!(action.subactions[1].internal_transitions, Javis.InternalTranslation(O))
+        Javis.compute_transformation!(action.subactions[1], video, 1)
+        @test action.subactions[1].internal_transitions[1].by == Point(1, 1)
+        Javis.compute_transformation!(action.subactions[1], video, 50)
+        @test action.subactions[1].internal_transitions[1].by ==
+              Point(at(anim, m), at(anim, m))
+        Javis.compute_transformation!(action.subactions[1], video, 100)
+        @test action.subactions[1].internal_transitions[1].by == Point(100, 100)
+    end
+
+    @testset "Relative frames" begin
+        video = Video(500, 500)
+        action = Action(Rel(10), (args...) -> 1, Translation(Point(1, 1), Point(100, 100)))
+        # dummy action doesn't need a real function
+        test_file = javis(
+            video,
+            [
+                Action(1:100, (args...) -> 1, Translation(Point(1, 1), Point(100, 100))),
+                action,
+            ],
+        )
+        @test Javis.get_frames(action) == 101:110
+        rm(test_file)
+    end
+
+    @testset "translation from origin" begin
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Translation(Point(99, 99)))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalTranslation(O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].by == O
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].by == Point(49, 49)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].by == Point(99, 99)
+
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Translation(99, 99))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalTranslation(O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].by == O
+        Javis.compute_transformation!(action, video, 50)
+        @test action.internal_transitions[1].by == Point(49, 49)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].by == Point(99, 99)
+    end
+
+    @testset "rotations" begin
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Rotation(2π))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalRotation(0.0, O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].angle == 0.0
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].angle == 2π
+
+        video = Video(500, 500)
+        # dummy action doesn't need a real function
+        action = Action(1:100, () -> 1, Rotation(2π, Point(2.0, 5.0)))
+        # needs internal translation as well
+        push!(action.internal_transitions, Javis.InternalRotation(0.0, O))
+
+        Javis.compute_transformation!(action, video, 1)
+        @test action.internal_transitions[1].angle == 0.0
+        @test action.internal_transitions[1].center == Point(2.0, 5.0)
+        Javis.compute_transformation!(action, video, 100)
+        @test action.internal_transitions[1].angle == 2π
+        @test action.internal_transitions[1].center == Point(2.0, 5.0)
     end
 
     @testset "scaling" begin