+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Javis.jl - Changelog
 
-## Unreleased
+## Unreleased v0.2
+- Ability to use [Animations.jl](https://github.com/jkrumbiegel/Animations.jl) 
+  - for Transformations and `appear` and `disappear`
+- Show progress of rendering using [ProgressMeter.jl](https://github.com/timholy/ProgressMeter.jl)
+- Use [VideoIO](https://github.com/JuliaIO/VideoIO.jl) for faster rendering without temporary images
 - Ability to scale an object with `Scaling`. Works similar to `Translation` and `Rotation` 
 - Added JuliaFormatter GitHub Action